requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-appconf==1.0.2
django-autoslug==1.9.4
django-avatar==4.1.0
django-heroku==0.3.1
django-storages==1.7.1
gunicorn==19.9.0
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
whitenoise==4.1.2
